fix: code cleanup for not necessary override methods
