+++ catalyst_voices/packages/internal/catalyst_voices_repositories/lib/src/utils/json_converters.dart
@@ -39,7 +39,7 @@ final class ListStringConverter
   @override
   List<String>? fromJson(List<dynamic>? json) {
     if (json == null) return null;
-    
+
     return json.cast();
   }
